+++ src/installNpmPkgs.js
@@ -4,8 +4,11 @@
  */
 const childProcess = require('child_process');
 const { npmPkgs } = require('./config');
+const chalk = require('chalk');
+const inquirer = require('inquirer');
+const prompt = inquirer.createPromptModule();
 
-module.exports = function ({ type }) {
+module.exports = async function ({ type }) {
   let npmCmd = 'npm i -D --legacy-peer-deps';
 
   const pkgTypes = ['eslint'];
@@ -14,6 +17,10 @@ module.exports = function ({ type }) {
     pkgTypes.push('prettier');
   }
   pkgTypes.push(type);
+
+  if (await checkTypescriptVersion(type)) {
+    pkgTypes.push('tsBase');
+  }
 
   for (const key of pkgTypes) {
     const pkgs = npmPkgs[key];
@@ -30,3 +37,27 @@ module.exports = function ({ type }) {
     if (err) return console.error('✘ Npm install failed', err);
   }
 };
+
+/** 提示用户是否安装typescript */
+async function checkTypescriptVersion(type) {
+  if (type.indexOf('ts') > -1) {
+    console.log('Checking typescript...');
+    // check user if install the typescript dep
+    const cmdReturn = childProcess.execSync('npm list typescript').toString();
+    const tsInfo = cmdReturn.match(/typescript@\d(\.\d{1,2}){2}/);
+    let defaultInstallTs = false;
+    if (tsInfo) {
+      console.log('Currently installed: ', chalk.green(tsInfo[0].toLowerCase()));
+    } else {
+      defaultInstallTs = true;
+      console.log('Typescript not currently installed.');
+    }
+    const answer = await prompt({
+      type: 'confirm',
+      name: 'installTs',
+      message: `Install typescript@${npmPkgs.tsBase.typescript}?`,
+      default: defaultInstallTs,
+    });
+    return answer.installTs;
+  }
+}