+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.7
-* add `'prefer-const': 1` in `js`,`vue2`,`vue3`
+* add `'prefer-const': 1` in `js`,`vue2`,`vue3`、
+* question npm install `typescript` when selected contain `ts` 
 ## 1.4.6
 * fix: select `ts` and operation error
 ## 1.4.5